refactor: move coder client under src/services/coder/
