+++ src/handlers/create-task.ts
@@ -1,5 +1,5 @@
-import type { CoderClient } from "../coder-client";
-import { TaskNameSchema } from "../coder-client";
+import type { CoderClient } from "../services/coder/client";
+import { TaskNameSchema } from "../services/coder/client";
 import type { GitHubClient } from "../github-client";
 import type { Logger } from "../infra/logger";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
