+++ src/handlers/failed-check.ts
@@ -1,4 +1,4 @@
-import type { CoderClient } from "../coder-client";
+import type { CoderClient } from "../services/coder/client";
 import type { GitHubClient, PRInfo } from "../github-client";
 import type { Logger } from "../infra/logger";
 import { MAX_FAILED_JOBS, formatFailedCheckMessage } from "../messages";
