+++ src/task-utils.ts
@@ -1,5 +1,8 @@
-import type { CoderClient, ExperimentalCoderSDKTask } from "./coder-client";
-import { CoderAPIError, TaskNameSchema } from "./coder-client";
+import type {
+	CoderClient,
+	ExperimentalCoderSDKTask,
+} from "./services/coder/client";
+import { CoderAPIError, TaskNameSchema } from "./services/coder/client";
 import type { Logger } from "./infra/logger";
 
 const DEFAULT_MAX_RETRIES = 5;
